refactor: Move common Platform code to `nextflow.platform.PlatformHelper`
